+++ app/src/main/java/com/group5/cmiopenday/QuestionFormActivity.java
@@ -1,20 +1,18 @@
 package com.group5.cmiopenday;
 
+import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
+import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Button;
 
-public class QuestionFormActivity extends MainActivity{
+public class QuestionFormActivity extends menu_Activity{
 
     private DrawerLayout drawer;
 
@@ -24,7 +22,9 @@ public class QuestionFormActivity extends MainActivity{
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
         drawer = findViewById(R.id.drawer_layout);
-        getLayoutInflater().inflate(R.layout.activity_questionform, drawer);
+        LayoutInflater inflater = (LayoutInflater) this.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        View ContentView = inflater.inflate(R.layout.activity_questionform,null,false);
+        drawer.addView(ContentView,0);
 
 
 
@@ -82,6 +82,9 @@ public class QuestionFormActivity extends MainActivity{
 
 
 }
+
+
+
