+++ app/src/main/java/com/group5/cmiopenday/menu_Activity.java
@@ -1,0 +1,77 @@
+package com.group5.cmiopenday;
+
+import android.app.Activity;
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.design.widget.NavigationView;
+import android.support.v4.view.GravityCompat;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBarDrawerToggle;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.view.MenuItem;
+
+
+public class menu_Activity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
+
+
+    private DrawerLayout drawer;
+    private NavigationView navigationView;
+
+    protected void onCreateDrawer(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        Toolbar toolbar = findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+        setContentView(R.layout.activity_main);
+
+
+
+        drawer = findViewById(R.id.drawer_layout);
+        ActionBarDrawerToggle toggle = new ActionBarDrawerToggle(this, drawer, toolbar, R.string.navigation_drawer_open, R.string.navigation_drawer_close);
+        drawer.addDrawerListener(toggle);
+        toggle.syncState();
+
+        navigationView = findViewById(R.id.sidebar_view);
+        navigationView.setNavigationItemSelectedListener(this);
+    }
+
+
+
+    @Override
+    public void onBackPressed() {
+        if (drawer.isDrawerOpen(GravityCompat.START)){
+            drawer.closeDrawer(GravityCompat.START);
+        }
+        else{
+            super.onBackPressed();
+        }
+    }
+
+
+    @Override
+    public boolean onNavigationItemSelected(MenuItem menuItem) {
+        int id = menuItem.getItemId();
+        switch (id) {
+            case R.id.sidebare_homescreen:
+                Intent a = new Intent(menu_Activity.this, MainActivity.class);
+                startActivity(a);
+                break;
+            case R.id.sidebare_location:
+                Intent b = new Intent(menu_Activity.this, EducationPage.class);
+                startActivity(b);
+                break;
+            case R.id.sidebare_study_programs:
+                Intent c = new Intent(menu_Activity.this, CMI_Activity.class);
+                startActivity(c);
+                break;
+            case R.id.sidebare_ask_a_question:
+                Intent d = new Intent(menu_Activity.this, QuestionFormActivity.class);
+                startActivity(d);
+                break;
+        }
+        DrawerLayout drawer = findViewById(R.id.drawer_layout);
+        drawer.closeDrawer(GravityCompat.START);
+        return true;
+    }
+
+}
